pyproject.toml
[tool.poetry]
name = "rag-toy"
version = "0.1.0"
description = "RAG Toy Project"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pdfplumber = "*"
numpy = "*"
pillow = "*"
pypdf = "*"
pytesseract = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
ruff = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

packages = [
  { include = "rag_toy", from = "src" }
]

[dependency-groups]
dev = [
    "azure-search-documents>=11.6.0",
    "bs4>=0.0.2",
    "fastapi>=0.128.0",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-azureaisearch>=0.4.4",
    "openai>=2.14.0",
]
